refactor: rename components(PascaCase), const file(kebab case)
