+++ src/components/ActionButtonGroup/index.module.scss
@@ -1,0 +1,15 @@
+.actionButtons {
+  .doneBtn {
+    width: 1.5rem;
+  }
+
+  .editBtn {
+    width: 1.5rem;
+    margin-left: 1.625rem;
+    margin-right: 1.813rem;
+  }
+
+  .deleteBtn {
+    width: 1.25rem;
+  }
+}
