+++ src/components/FilterButtonGroup/index.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Button from "components/common/button";
+import Button from "components/Common/Button";
 import {
   TASK_FILTER_ALL,
   TASK_FILTER_COMPLETED,
