+++ src/components/common/TextArea/index.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useRef } from "react";
 import PropTypes from "prop-types";
-import { TEXTAREA_DEFAULT_ROW } from "utils/const/formElements";
+import { TEXTAREA_DEFAULT_ROW } from "utils/const/form-elements";
 
 function TextArea(props) {
   const { errorMessage, numOfRows, name, autoFocus, value, onChange } = props;
