+++ src/components/common/button/index.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import PropTypes from "prop-types";
 import classNames from "classnames";
-import { BUTTON_TYPE_SUBMIT } from "utils/const/formElements";
-import styles from "components/common/button/index.module.scss";
+import { BUTTON_TYPE_SUBMIT } from "utils/const/form-elements";
+import styles from "components/Common/Button/index.module.scss";
 
 function Button({ buttonType, children, className, onClick }) {
   const buttonClass = classNames(styles.button, className);
