--- src/components/task/existing-task/action-button-group/index.module.scss
@@ -1,13 +1,0 @@
-.doneBtn {
-  width: 1.5rem;
-}
-
-.editBtn {
-  width: 1.5rem;
-  margin-left: 1.625rem;
-  margin-right: 1.813rem;
-}
-
-.deleteBtn {
-  width: 1.25rem;
-}
