+++ src/utils/factory.js
@@ -1,0 +1,9 @@
+export const generateTaskObject = (title) => {
+  return {
+    id: Date.now().toString(),
+    title: title,
+    createdAt: new Date(),
+    isCompleted: false,
+    completedAt: null,
+  };
+};
