+++ src/utils/helper/index.js
@@ -1,13 +1,3 @@
-export const generateTaskObject = (title) => {
-  return {
-    id: Date.now().toString(),
-    title: title,
-    createdAt: new Date(),
-    isCompleted: false,
-    completedAt: null,
-  };
-};
-
 export const formatDate = (date) => {
   const today = new Date(date);
   const day = today.getDate();
